Attempt to set attribute value without using Gameplay Effects
The demo code crashes the game with an access code violation
so the engine must have changed something that inadvertently
makes it more difficult--ironically hightlighting why it is bad
practice to cast_const to get around value immutability.

I didn't commit the code that attempts to set an attribute value
but left the cast to AuraAttributeSet for future reference.

Create a health potion blueprint and place it in the scene.

Overlapping with the potion simply destroys it.